feat: header ui 개선
