+++ frontend/src/components/common/ActionButton.tsx
@@ -4,7 +4,7 @@ const ActionButton = ({
   disabled,
   children,
 }: {
-  onClick: () => void;
+  onClick: () => void | Promise<void>;
   color: string;
   disabled?: boolean;
   children: React.ReactNode;
@@ -12,7 +12,7 @@ const ActionButton = ({
   return (
     <button
       onClick={onClick}
-      className={`${color} text-white rounded-lg px-4 py-2 flex items-center justify-center gap-1 cursor-pointer`}
+      className={`${color} text-white rounded-lg px-4 py-2 flex items-center justify-center gap-1 cursor-pointer disabled:opacity-50 disabled:cursor-not-allowed`}
       disabled={disabled}
     >
       {children}
